refactor monopoly code
