+++ honeybot/plugins/monopoly_assets/monopoly_assets.py
